keys: support account compilation
