+++ keys/Cargo.toml
@@ -10,4 +10,5 @@ edition = { workspace = true }
 version = { workspace = true }
 
 [dependencies]
+mollusk-svm-error = { workspace = true }
 solana-sdk = { workspace = true }
