+++ keys/src/lib.rs
@@ -1,3 +1,4 @@
 //! SVM transaction keys utils.
 
+pub mod accounts;
 pub mod keys;
